Cleans up logic & dependencies for requesting sf/sp objects.
This commit moves the `sf` package to the Suggests section of the DESCRIPTION
file, since it is possible to use the package entirely with the `sp` stack. The
package seems to have supported GeoJSON for quite some time, so I've removed the
requirement for the most recent version (0.5-3).

The `rdgal` package is now mentioned in the DESCRIPTION file as well. It was
previously absent.