Use Imports, not Depends, in package DESCRIPTION file.
There is no need for this package to attach all of its dependencies, so we
should be using Imports.

This commit also changes the minimum required versions for the imported
packages, because we don't really need to make users install the most recent
version.

- `digest` has supported the MD5 hashing format since its inception, at least
  according to its CHANGELOG file.
- `jsonlite` has also supported `toJSON()` for quite some time, so adopt its
  first "complete" release, version 1.0.
- `httr` has a similar story, so adopt its version 1.0.0 as well.

I've not looked into the minimum version of `dplyr` required.