Fix vignette so it can compile without additional packages
The errors in compiling the package was due to this vignette requiring libraries that are not dependencies to the overall package. Specifically, it's ggthemes, which also requires mapproj. I don't think it's necessary for the visual intended in the vignette and I cut it out, and made a small graphical change to the legend as well as it was buggy when testing otherwise.